requirements.txt
cryptography>=41.0.0
pycryptodome>=3.19.0
dearpygui>=2.1.0
Pillow>=10.0.0